+++ fastai/callback.py
@@ -243,7 +243,7 @@ class CallbackHandler():
 
     def on_epoch_end(self, val_loss:Tensor)->bool:
         "Epoch is done, process `val_metrics`."
-        self.state_dict['last_metrics'] = val_loss
+        self.state_dict['last_metrics'] = [val_loss] if val_loss is not None else None
         self.state_dict['epoch'] += 1
         if not self.state_dict['train']:
             for met in self.metrics: