+++ fastai/callbacks/hooks.py
@@ -9,13 +9,13 @@ __all__ = ['ActivationStats', 'Hook', 'HookCallback', 'Hooks', 'hook_output', 'h
 
 class Hook():
     "Create a hook."
-    def __init__(self, m:Model, hook_func:HookFunc, is_forward:bool=True):
+    def __init__(self, m:nn.Module, hook_func:HookFunc, is_forward:bool=True):
         self.hook_func,self.stored = hook_func,None
         f = m.register_forward_hook if is_forward else m.register_backward_hook
         self.hook = f(self.hook_fn)
         self.removed = False
 
-    def hook_fn(self, module:Model, input:Tensors, output:Tensors):
+    def hook_fn(self, module:nn.Module, input:Tensors, output:Tensors):
         input  = (o.detach() for o in input ) if is_listy(input ) else input.detach()
         output = (o.detach() for o in output) if is_listy(output) else output.detach()
         self.stored = self.hook_func(module, input, output)
@@ -27,7 +27,7 @@ class Hook():
 
 class Hooks():
     "Create several hooks."
-    def __init__(self, ms:Collection[Model], hook_func:HookFunc, is_forward:bool=True):
+    def __init__(self, ms:Collection[nn.Module], hook_func:HookFunc, is_forward:bool=True):
         self.hooks = [Hook(m, hook_func, is_forward) for m in ms]
 
     def __getitem__(self,i:int) -> Hook: return self.hooks[i]
@@ -39,12 +39,12 @@ class Hooks():
     def remove(self):
         for h in self.hooks: h.remove()
 
-def hook_output (module:Model) -> Hook:  return Hook (module,  lambda m,i,o: o)
-def hook_outputs(modules:Collection[Model]) -> Hooks: return Hooks(modules, lambda m,i,o: o)
+def hook_output (module:nn.Module) -> Hook:  return Hook (module,  lambda m,i,o: o)
+def hook_outputs(modules:Collection[nn.Module]) -> Hooks: return Hooks(modules, lambda m,i,o: o)
 
 class HookCallback(LearnerCallback):
     "Callback that registers given hooks."
-    def __init__(self, learn:Learner, modules:Sequence[Model]=None, do_remove:bool=True):
+    def __init__(self, learn:Learner, modules:Sequence[nn.Module]=None, do_remove:bool=True):
         super().__init__(learn)
         self.modules,self.do_remove = modules,do_remove
 
@@ -66,17 +66,17 @@ class ActivationStats(HookCallback):
         super().on_train_begin(**kwargs)
         self.stats = []
 
-    def hook(self, m:Model, i:Tensors, o:Tensors) -> Tuple[Rank0Tensor,Rank0Tensor]:
+    def hook(self, m:nn.Module, i:Tensors, o:Tensors) -> Tuple[Rank0Tensor,Rank0Tensor]:
         return o.mean().item(),o.std().item()
     def on_batch_end(self, train, **kwargs): 
         if train: self.stats.append(self.hooks.stored)
     def on_train_end(self, **kwargs): self.stats = tensor(self.stats).permute(2,1,0)
 
-def model_sizes(m:Model, size:tuple=(256,256), full:bool=True) -> Tuple[Sizes,Tensor,Hooks]:
+def model_sizes(m:nn.Module, size:tuple=(256,256), full:bool=True) -> Tuple[Sizes,Tensor,Hooks]:
     "Pass a dummy input through the model to get the various sizes."
     hooks = hook_outputs(m)
     ch_in = in_channels(m)
-    x = torch.zeros(1,ch_in,*size)
+    x = next(m.parameters()).new(1,ch_in,*size)
     x = m.eval()(x)
     res = [o.stored.shape for o in hooks]
     if not full: hooks.remove()