+++ fastai/vision/learner.py
@@ -7,17 +7,28 @@ from . import models
 from ..callback import *
 from ..layers import *
 
-__all__ = ['ConvLearner', 'create_body', 'create_head', 'num_features', 'ClassificationInterpretation']
+__all__ = ['create_cnn', 'create_body', 'create_head', 'ClassificationInterpretation']
+# By default split models between first and second layer
+def _default_split(m:nn.Module): return (m[1],)
+# Split a resnet style model
+def _resnet_split(m:nn.Module): return (m[0][6],m[1])
 
-def create_body(model:Model, cut:Optional[int]=None, body_fn:Callable[[Model],Model]=None):
+_default_meta = {'cut':-1, 'split':_default_split}
+_resnet_meta  = {'cut':-2, 'split':_resnet_split }
+
+model_meta = {
+    models.resnet18 :{**_resnet_meta}, models.resnet34: {**_resnet_meta},
+    models.resnet50 :{**_resnet_meta}, models.resnet101:{**_resnet_meta},
+    models.resnet152:{**_resnet_meta}}
+
+def cnn_config(arch):
+    torch.backends.cudnn.benchmark = True
+    return model_meta.get(arch, _default_meta)
+
+def create_body(model:nn.Module, cut:Optional[int]=None, body_fn:Callable[[nn.Module],nn.Module]=None):
     "Cut off the body of a typically pretrained `model` at `cut` or as specified by `body_fn`."
     return (nn.Sequential(*list(model.children())[:cut]) if cut
             else body_fn(model) if body_fn else model)
-
-def num_features(m:Model)->int:
-    "Return the number of output features for a `model`."
-    for l in reversed(flatten_model(m)):
-        if hasattr(l, 'num_features'): return l.num_features
 
 def create_head(nf:int, nc:int, lin_ftrs:Optional[Collection[int]]=None, ps:Floats=0.5):
     """Model head that takes `nf` features, runs through `lin_ftrs`, and about `nc` classes.
@@ -31,49 +42,35 @@ def create_head(nf:int, nc:int, lin_ftrs:Optional[Collection[int]]=None, ps:Floa
         layers += bn_drop_lin(ni,no,True,p,actn)
     return nn.Sequential(*layers)
 
-# By default split models between first and second layer
-def _default_split(m:Model): return (m[1],)
-# Split a resnet style model
-def _resnet_split(m:Model): return (m[0][6],m[1])
 
-_default_meta = {'cut':-1, 'split':_default_split}
-_resnet_meta  = {'cut':-2, 'split':_resnet_split }
-
-model_meta = {
-    models.resnet18 :{**_resnet_meta}, models.resnet34: {**_resnet_meta},
-    models.resnet50 :{**_resnet_meta}, models.resnet101:{**_resnet_meta},
-    models.resnet152:{**_resnet_meta}}
-
-class ConvLearner(Learner):
+def create_cnn(data:DataBunch, arch:Callable, cut:Union[int,Callable]=None, pretrained:bool=True,
+                lin_ftrs:Optional[Collection[int]]=None, ps:Floats=0.5,
+                custom_head:Optional[nn.Module]=None, split_on:Optional[SplitFuncOrIdxList]=None, **kwargs:Any)->None:
     "Build convnet style learners."
-    def __init__(self, data:DataBunch, arch:Callable, cut:Union[int,Callable]=None, pretrained:bool=True,
-                 lin_ftrs:Optional[Collection[int]]=None, ps:Floats=0.5,
-                 custom_head:Optional[nn.Module]=None, split_on:Optional[SplitFuncOrIdxList]=None, **kwargs:Any)->None:
-        meta = model_meta.get(arch, _default_meta)
-        torch.backends.cudnn.benchmark = True
-        body = create_body(arch(pretrained), ifnone(cut,meta['cut']))
-        nf = num_features(body) * 2
-        head = custom_head or create_head(nf, data.c, lin_ftrs, ps)
-        model = nn.Sequential(body, head)
-        super().__init__(data, model, **kwargs)
-        self.split(ifnone(split_on,meta['split']))
-        if pretrained: self.freeze()
-        apply_init(model[1], nn.init.kaiming_normal_)
+    meta = cnn_config(arch)
+    body = create_body(arch(pretrained), ifnone(cut,meta['cut']))
+    nf = num_features_model(body) * 2
+    head = custom_head or create_head(nf, data.c, lin_ftrs, ps)
+    model = nn.Sequential(body, head)
+    learn = Learner(data, model, **kwargs)
+    learn.split(ifnone(split_on,meta['split']))
+    if pretrained: learn.freeze()
+    apply_init(model[1], nn.init.kaiming_normal_)
+    return learn
+
 
 class ClassificationInterpretation():
     "Interpretation methods for classification models."
-    def __init__(self, data:DataBunch, y_pred:Tensor, y_true:Tensor,
-                 loss_class:type=nn.CrossEntropyLoss, sigmoid:bool=True):
-        self.data,self.y_pred,self.y_true,self.loss_class = data,y_pred,y_true,loss_class
-        self.losses = calc_loss(y_pred, y_true, loss_class=loss_class)
-        self.probs = y_pred.sigmoid() if sigmoid else y_pred
+    def __init__(self, data:DataBunch, probs:Tensor, y_true:Tensor, losses:Tensor, sigmoid:bool=None):
+        if sigmoid is not None: warnings.warn("`sigmoid` argument is deprecated, the learner now always return the probabilities")
+        self.data,self.probs,self.y_true,self.losses = data,probs,y_true,losses
         self.pred_class = self.probs.argmax(dim=1)
 
     @classmethod
-    def from_learner(cls, learn:Learner, loss_class:type=nn.CrossEntropyLoss, sigmoid:bool=True, tta=False):
+    def from_learner(cls, learn:Learner, sigmoid:bool=None, tta=False):
         "Factory method to create from a Learner."
-        preds = learn.TTA() if tta else learn.get_preds()
-        return cls(learn.data, *preds, loss_class=loss_class, sigmoid=sigmoid)
+        preds = learn.TTA(with_loss=True) if tta else learn.get_preds(with_loss=True)
+        return cls(learn.data, *preds, sigmoid=sigmoid)
 
     def top_losses(self, k, largest=True):
         "`k` largest(/smallest) losses."
@@ -89,6 +86,7 @@ class ClassificationInterpretation():
         classes = self.data.classes
         rows = math.ceil(math.sqrt(k))
         fig,axes = plt.subplots(rows,rows,figsize=figsize)
+        fig.suptitle('prediction/actual/loss/probability', weight='bold', size=14)
         for i,idx in enumerate(tl_idx):
             t=self.data.valid_ds[idx]
             t[0].show(ax=axes.flat[i], title=
@@ -128,11 +126,12 @@ class ClassificationInterpretation():
                 for i,j in zip(*np.where(cm>min_val))]
         return sorted(res, key=itemgetter(2), reverse=True)
 
-def _predict(img, learn):
+def Image_predict(img, learn):
     img = apply_tfms(learn.data.valid_ds.tfms, img, **learn.data.valid_ds.kwargs)
     ds = TensorDataset(img.data[None], torch.zeros(1))
     dl = DeviceDataLoader.create(ds, bs=1, shuffle=False, device=learn.data.device, tfms=learn.data.valid_dl.tfms,
                                  num_workers=0)
     return get_preds(learn.model, dl, cb_handler=CallbackHandler(learn.callbacks, []))[0][0]
 
-Image.predict = _predict
+Image.predict = Image_predict
+