+++ tests/test_core.py
@@ -1,6 +1,7 @@
 import pytest, torch
 import numpy as np
 from fastai import *
+from tempfile import TemporaryDirectory
 
 def test_cpus(): assert num_cpus() >= 1
 
@@ -145,3 +146,44 @@ def test_series2cat():
     for col in cols:
         assert (df[col].dtypes == 'category')
     assert (df['col3'].dtypes == 'int64')
+
+def _write_file(path): f = open(path, 'w'); f.write(str(path.name)); f.close()
+class TestMaybeCopy(object):
+    def test_copies_if_does_not_exist(self):
+        with TemporaryDirectory() as tmpdir:
+            tmpdir = Path(tmpdir)
+            _write_file(tmpdir/'src')
+            maybe_copy([str(tmpdir/'src')], [str(tmpdir/'dst')]) # works with strings
+            assert os.path.exists(tmpdir/'dst')
+        with TemporaryDirectory() as tmpdir:
+            tmpdir = Path(tmpdir)
+            _write_file(tmpdir/'src')
+            maybe_copy([tmpdir/'src'], [tmpdir/'dst']) # works with Paths
+            assert os.path.exists(tmpdir/'dst')
+
+    def test_copies_if_older(self):
+        with TemporaryDirectory() as tmpdir:
+            tmpdir = Path(tmpdir)
+            _write_file(tmpdir/'first')
+            _write_file(tmpdir/'second')
+            os.utime(tmpdir/'first', (1,1))
+            os.utime(tmpdir/'second', (2,2))
+            maybe_copy([tmpdir/'second'], [tmpdir/'first'])
+            assert open(tmpdir/'first').read() == 'second'
+
+    def test_does_not_copy_if_newer(self):
+        with TemporaryDirectory() as tmpdir:
+            tmpdir = Path(tmpdir)
+            _write_file(tmpdir/'first')
+            _write_file(tmpdir/'second')
+            os.utime(tmpdir/'first', (1,1))
+            os.utime(tmpdir/'second', (2,2))
+            maybe_copy([tmpdir/'first'], [tmpdir/'second'])
+            assert open(tmpdir/'second').read() == 'second'
+
+    def test_creates_dst_dir_if_does_not_exist(self):
+        with TemporaryDirectory() as tmpdir:
+            tmpdir = Path(tmpdir)
+            _write_file(tmpdir/'file')
+            maybe_copy([tmpdir/'file'], [tmpdir/'dir'/'file'])
+            assert os.path.exists(tmpdir/'dir'/'file')