+++ tests/test_text_train.py
@@ -48,7 +48,6 @@ def test_classifier():
             df = text_df(n_labels=n_labels)
             data = TextClasDataBunch.from_df(path, train_df=df, valid_df=df, label_cols=list(range(n_labels)), txt_cols=["text"])
             classifier = RNNLearner.classifier(data)
-            layers = [l for l in classifier.layer_groups[-1].modules()]
-            assert layers[-1].out_features == n_labels if n_labels > 1 else n_labels+1
+            assert last_layer(classifier.model).out_features == n_labels if n_labels > 1 else n_labels+1
         finally:
             shutil.rmtree(path)